{"schema":"decodex/commit/1","summary":"Split Swift app bridge networking modules by responsibility","authority":"manual"}
